+++ src/asset.rs
@@ -7,7 +7,7 @@ use crate::util::{build_json_request, build_request};
 #[derive(Clone)]
 pub struct AssetManager {
     pub client: Client,
-    pub recv_window: u16,
+    pub recv_window: u64,
 }
 
 impl AssetManager {
